fix: 이메일, 닉네임 검증 및 중복 확인 메소드 AuthService로 이동
