+++ src/main/java/kr/co/fastcampus/yanabada/domain/auth/service/AuthService.java
@@ -11,10 +11,16 @@ import kr.co.fastcampus.yanabada.common.jwt.dto.TokenRefreshResponse;
 import kr.co.fastcampus.yanabada.common.jwt.service.TokenService;
 import kr.co.fastcampus.yanabada.common.jwt.util.JwtProvider;
 import kr.co.fastcampus.yanabada.common.utils.CookieCreator;
+import kr.co.fastcampus.yanabada.domain.auth.dto.request.AuthCodeVerifyRequest;
+import kr.co.fastcampus.yanabada.domain.auth.dto.request.EmailCodeSendRequest;
 import kr.co.fastcampus.yanabada.domain.auth.dto.request.LoginRequest;
 import kr.co.fastcampus.yanabada.domain.auth.dto.request.OauthSignUpRequest;
 import kr.co.fastcampus.yanabada.domain.auth.dto.request.SignUpRequest;
+import kr.co.fastcampus.yanabada.domain.auth.dto.response.AuthCodeVerifyResponse;
 import kr.co.fastcampus.yanabada.domain.auth.dto.response.SignUpResponse;
+import kr.co.fastcampus.yanabada.domain.member.dto.request.EmailDuplCheckRequest;
+import kr.co.fastcampus.yanabada.domain.member.dto.request.NickNameDuplCheckRequest;
+import kr.co.fastcampus.yanabada.domain.member.dto.response.DuplCheckResponse;
 import kr.co.fastcampus.yanabada.domain.member.entity.Member;
 import kr.co.fastcampus.yanabada.domain.member.entity.ProviderType;
 import kr.co.fastcampus.yanabada.domain.member.repository.MemberRepository;
@@ -51,6 +57,7 @@ public class AuthService {
     private final JwtProvider jwtProvider;
     private final AuthenticationManagerBuilder authenticationManagerBuilder;
     private final TokenService tokenService;
+    private final MailAuthService mailAuthService;
 
     @Transactional
     public SignUpResponse signUp(SignUpRequest signUpRequest) {
@@ -154,5 +161,42 @@ public class AuthService {
         String newAccessToken = jwtProvider.generateAccessToken(email, role, provider);
         tokenService.updateAccessToken(email, provider, newAccessToken);
         return new TokenRefreshResponse(newAccessToken);
+    }
+
+
+    @Transactional(readOnly = true)
+    public void sendAuthCodeToEmail(
+        EmailCodeSendRequest emailRequest
+    ) {
+        boolean isExist = memberRepository
+            .existsByEmailAndProviderType(emailRequest.email(), EMAIL);
+        if (isExist) {
+            throw new EmailDuplicatedException();
+        }
+        mailAuthService.sendEmail(emailRequest.email());
+    }
+
+    @Transactional(readOnly = true)
+    public AuthCodeVerifyResponse verifyAuthCode(AuthCodeVerifyRequest codeRequest) {
+        return new AuthCodeVerifyResponse(
+            mailAuthService.verifyAuthCode(codeRequest.email(), codeRequest.code())
+        );
+    }
+
+    @Transactional(readOnly = true)
+    public DuplCheckResponse isExistEmail(
+        EmailDuplCheckRequest emailRequest
+    ) {
+        boolean isExist = memberRepository
+            .existsByEmailAndProviderType(emailRequest.email(), EMAIL);
+        return new DuplCheckResponse(isExist);
+    }
+
+    @Transactional(readOnly = true)
+    public DuplCheckResponse isExistNickName(
+        NickNameDuplCheckRequest nickNameRequest
+    ) {
+        boolean isExist = memberRepository.existsByNickName(nickNameRequest.nickName());
+        return new DuplCheckResponse(isExist);
     }
 }
