+++ src/main/java/kr/co/fastcampus/yanabada/domain/member/service/MemberService.java
@@ -1,20 +1,9 @@
 package kr.co.fastcampus.yanabada.domain.member.service;
 
-import static kr.co.fastcampus.yanabada.domain.member.entity.ProviderType.EMAIL;
-
-import kr.co.fastcampus.yanabada.common.exception.EmailDuplicatedException;
-import kr.co.fastcampus.yanabada.common.redis.RedisUtils;
-import kr.co.fastcampus.yanabada.domain.auth.dto.request.AuthCodeVerifyRequest;
-import kr.co.fastcampus.yanabada.domain.auth.dto.request.EmailCodeSendRequest;
-import kr.co.fastcampus.yanabada.domain.auth.dto.response.AuthCodeVerifyResponse;
-import kr.co.fastcampus.yanabada.domain.auth.service.MailAuthService;
-import kr.co.fastcampus.yanabada.domain.member.dto.request.EmailDuplCheckRequest;
 import kr.co.fastcampus.yanabada.domain.member.dto.request.FcmTokenUpdateRequest;
-import kr.co.fastcampus.yanabada.domain.member.dto.request.NickNameDuplCheckRequest;
 import kr.co.fastcampus.yanabada.domain.member.dto.request.NickNameModifyRequest;
 import kr.co.fastcampus.yanabada.domain.member.dto.request.PasswordModifyRequest;
 import kr.co.fastcampus.yanabada.domain.member.dto.request.PhoneNumberModifyRequest;
-import kr.co.fastcampus.yanabada.domain.member.dto.response.DuplCheckResponse;
 import kr.co.fastcampus.yanabada.domain.member.dto.response.MemberDetailResponse;
 import kr.co.fastcampus.yanabada.domain.member.entity.Member;
 import kr.co.fastcampus.yanabada.domain.member.entity.ProviderType;
@@ -32,7 +21,6 @@ public class MemberService {
 
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
-    private final MailAuthService mailAuthService;
 
     @Transactional(readOnly = true)
     public MemberDetailResponse findMember(String email, ProviderType providerType) {
@@ -67,42 +55,6 @@ public class MemberService {
     ) {
         Member member = memberRepository.getMember(email, providerType);
         member.updatePhoneNumber(phoneNumberRequest.phoneNumber());
-    }
-
-    @Transactional(readOnly = true)
-    public void sendAuthCodeToEmail(
-        EmailCodeSendRequest emailRequest
-    ) {
-        boolean isExist = memberRepository
-            .existsByEmailAndProviderType(emailRequest.email(), EMAIL);
-        if (isExist) {
-            throw new EmailDuplicatedException();
-        }
-        mailAuthService.sendEmail(emailRequest.email());
-    }
-
-    @Transactional(readOnly = true)
-    public AuthCodeVerifyResponse verifyAuthCode(AuthCodeVerifyRequest codeRequest) {
-        return new AuthCodeVerifyResponse(
-            mailAuthService.verifyAuthCode(codeRequest.email(), codeRequest.code())
-        );
-    }
-
-    @Transactional(readOnly = true)
-    public DuplCheckResponse isExistEmail(
-        EmailDuplCheckRequest emailRequest
-    ) {
-        boolean isExist = memberRepository
-            .existsByEmailAndProviderType(emailRequest.email(), EMAIL);
-        return new DuplCheckResponse(isExist);
-    }
-
-    @Transactional(readOnly = true)
-    public DuplCheckResponse isExistNickName(
-        NickNameDuplCheckRequest nickNameRequest
-    ) {
-        boolean isExist = memberRepository.existsByNickName(nickNameRequest.nickName());
-        return new DuplCheckResponse(isExist);
     }
 
     @Transactional
